fixed bugs found by QA. thanks David

+++ src/Catalyte.Apparel.DTOs/Promo/CreatePromoDTO.cs
@@ -9,7 +9,7 @@ namespace Catalyte.Apparel.DTOs.Promos
     {
         public string Code { get; set; }
 
-        public int? Discount { get; set; }
+        public decimal? Discount { get; set; }
 
         public string Type { get; set; }
